#2, Code for the Company Controller is done

+++ Job_App/src/main/java/com/Praveen/Job_App/Job/JobService.java
@@ -24,6 +24,9 @@ public class JobService {
     }
 
     public boolean deleteById(Long id) {
+
+        if(!repo.existsById(id)) return false;
+
         try{
             repo.deleteById(id);
             return true;
